+++ modules/core/src/model/datepicker/dynamic-datepicker.model.ts
@@ -1,0 +1,28 @@
+import { ClsConfig } from "../dynamic-form-control.model";
+import { serializable } from "../../decorator/serializable.decorator";
+import { getValue } from "../../utils";
+import { DynamicDateControlModel, DynamicDateControlModelConfig } from "../dynamic-date-control.model";
+
+export const DYNAMIC_FORM_CONTROL_TYPE_DATEPICKER = "DATEPICKER";
+
+export interface DynamicDatepickerModelConfig extends DynamicDateControlModelConfig {
+
+    focusedDate?: Date;
+    inline?: boolean;
+}
+
+export class DynamicDatepickerModel extends DynamicDateControlModel {
+
+    @serializable() focusedDate: Date | null;
+    @serializable() inline: boolean;
+
+    @serializable() readonly type: string = DYNAMIC_FORM_CONTROL_TYPE_DATEPICKER;
+
+    constructor(config: DynamicDatepickerModelConfig, cls?: ClsConfig) {
+
+        super(config, cls);
+
+        this.focusedDate = getValue(config, "focusedDate", null);
+        this.inline = getValue(config, "inline", false);
+    }
+}