+++ modules/core/src/service/dynamic-form.service.ts
@@ -19,6 +19,10 @@ import {
     DynamicCheckboxGroupModel
 } from "../model/checkbox/dynamic-checkbox-group.model";
 import { DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX, DynamicCheckboxModel } from "../model/checkbox/dynamic-checkbox.model";
+import {
+    DYNAMIC_FORM_CONTROL_TYPE_DATEPICKER,
+    DynamicDatepickerModel
+} from "../model/datepicker/dynamic-datepicker.model";
 import { DYNAMIC_FORM_CONTROL_TYPE_INPUT, DynamicInputModel } from "../model/input/dynamic-input.model";
 import {
     DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP,
@@ -33,8 +37,17 @@ import { isFunction, isDefined } from "../utils";
 export class DynamicFormService {
 
     constructor(@Inject(FormBuilder) private formBuilder: FormBuilder,
-                @Optional() @Inject(NG_VALIDATORS) private NG_VALIDATORS: Array<ValidatorFn>,
-                @Optional() @Inject(NG_ASYNC_VALIDATORS) private NG_ASYNC_VALIDATORS: Array<AsyncValidatorFn>) {}
+                @Optional() @Inject(NG_VALIDATORS) private NG_VALIDATORS: ValidatorFn[],
+                @Optional() @Inject(NG_ASYNC_VALIDATORS) private NG_ASYNC_VALIDATORS: AsyncValidatorFn[]) {}
+
+
+    private parseJSONReviver(key: string, value: any): any {
+
+        let regexISO = /^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2}(?:\.\d*))(?:Z|(\+|-)([\d|:]*))?$/;
+
+        return typeof value === "string" && regexISO.test(value) ? new Date(value) : value;
+    }
+
 
     getCustomValidatorFn(validatorName: string): ValidatorFn | AsyncValidatorFn | undefined {
 
@@ -51,6 +64,7 @@ export class DynamicFormService {
         return validatorFn;
     }
 
+
     getValidatorFn(validatorName: string, validatorArgs?: any): ValidatorFn | AsyncValidatorFn | never {
 
         let validatorFn = Validators[validatorName] || this.getCustomValidatorFn(validatorName);
@@ -62,11 +76,13 @@ export class DynamicFormService {
         return isDefined(validatorArgs) ? validatorFn(validatorArgs) : validatorFn;
     }
 
-    getValidators(config: DynamicValidatorsMap): Array<ValidatorFn | AsyncValidatorFn> {
+
+    getValidators(config: DynamicValidatorsMap): ValidatorFn[] | AsyncValidatorFn[] {
 
         return isDefined(config) ?
             Object.keys(config).map(validatorName => this.getValidatorFn(validatorName, config[validatorName])) : [];
     }
+
 
     createFormArray(model: DynamicFormArrayModel): FormArray {
 
@@ -83,7 +99,8 @@ export class DynamicFormService {
         );
     }
 
-    createFormGroup(group: Array<DynamicFormControlModel>, groupExtra: {[key: string]: any} | null = null): FormGroup {
+
+    createFormGroup(group: DynamicFormControlModel[], groupExtra: {[key: string]: any} | null = null): FormGroup {
 
         let formGroup = {};
 
@@ -91,13 +108,13 @@ export class DynamicFormService {
 
             if (model.type === DYNAMIC_FORM_CONTROL_TYPE_ARRAY) {
 
-                let arrayModel = <DynamicFormArrayModel> model;
+                let arrayModel = model as DynamicFormArrayModel;
 
                 formGroup[model.id] = this.createFormArray(arrayModel);
 
             } else if (model.type === DYNAMIC_FORM_CONTROL_TYPE_GROUP || model.type === DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP) {
 
-                let groupModel = <DynamicFormGroupModel> model,
+                let groupModel = model as DynamicFormGroupModel,
                     groupExtra = {
                         validator: this.getValidators(groupModel.validator)[0] || null,
                         asyncValidator: this.getValidators(groupModel.asyncValidator)[0] || null
@@ -107,7 +124,7 @@ export class DynamicFormService {
 
             } else {
 
-                let controlModel = <DynamicFormValueControlModel<DynamicFormControlValue>> model;
+                let controlModel = model as DynamicFormValueControlModel<DynamicFormControlValue>;
 
                 formGroup[controlModel.id] = new FormControl(
                     {
@@ -123,9 +140,10 @@ export class DynamicFormService {
         return this.formBuilder.group(formGroup, groupExtra);
     }
 
+
     addFormGroupControl(formGroup: FormGroup,
-                        groupModel: Array<DynamicFormControlModel> | DynamicFormGroupModel,
-                        ...controlModels: Array<DynamicFormControlModel>): void {
+                        groupModel: DynamicFormControlModel[] | DynamicFormGroupModel,
+                        ...controlModels: DynamicFormControlModel[]): void {
 
         let controls = this.createFormGroup(controlModels).controls;
 
@@ -137,16 +155,17 @@ export class DynamicFormService {
                 groupModel.add(controlModel);
 
             } else {
-                (<Array<DynamicFormControlModel>> groupModel).push(controlModel);
+                (groupModel as DynamicFormControlModel[]).push(controlModel);
             }
 
             formGroup.addControl(controlName, controls[controlName]);
         });
     }
 
+
     removeFormGroupControl(index: number,
                            formGroup: FormGroup,
-                           groupModel: Array<DynamicFormControlModel> | DynamicFormGroupModel) {
+                           groupModel: DynamicFormControlModel[] | DynamicFormGroupModel) {
 
         if (groupModel instanceof DynamicFormGroupModel) {
 
@@ -154,26 +173,31 @@ export class DynamicFormService {
             groupModel.remove(index);
 
         } else {
+
             formGroup.removeControl(groupModel[index].id);
-            (<Array<DynamicFormControlModel>> groupModel).splice(index, 1);
+            (groupModel as DynamicFormControlModel[]).splice(index, 1);
         }
     }
+
 
     addFormArrayGroup(formArray: FormArray, model: DynamicFormArrayModel): void {
 
         formArray.push(this.createFormGroup(model.addGroup().group));
     }
 
+
     insertFormArrayGroup(index: number, formArray: FormArray, model: DynamicFormArrayModel): void {
 
         formArray.insert(index, this.createFormGroup(model.insertGroup(index).group));
     }
+
 
     removeFormArrayGroup(index: number, formArray: FormArray, model: DynamicFormArrayModel): void {
 
         formArray.removeAt(index);
         model.removeGroup(index);
     }
+
 
     moveFormArrayGroup(index: number, step: number, formArray: FormArray, model: DynamicFormArrayModel): void {
 
@@ -191,6 +215,7 @@ export class DynamicFormService {
         }
     }
 
+
     clearFormArray(formArray: FormArray, model: DynamicFormArrayModel): void {
 
         while (formArray.length > 0) {
@@ -198,16 +223,19 @@ export class DynamicFormService {
         }
     }
 
-    findById(id: string, groupModel: Array<DynamicFormControlModel>): DynamicFormControlModel | undefined {
+
+    findById(id: string, groupModel: DynamicFormControlModel[]): DynamicFormControlModel | undefined {
 
         return groupModel.find(controlModel => controlModel.id === id);
     }
 
-    fromJSON(json: Array<Object>): Array<DynamicFormControlModel> | never {
 
-        let formModel: Array<DynamicFormControlModel> = [];
+    fromJSON(json: string | Object[]): DynamicFormControlModel[] | never {
 
-        json.forEach(model => {
+        let rawModel = typeof json === "string" ? JSON.parse(json, this.parseJSONReviver) : json,
+            formModel = [];
+
+        rawModel.forEach(model => {
 
             switch (model["type"]) {
 
@@ -224,6 +252,10 @@ export class DynamicFormService {
                 case DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP:
                     model["group"] = this.fromJSON(model["group"]);
                     formModel.push(new DynamicCheckboxGroupModel(model, model["cls"]));
+                    break;
+
+                case DYNAMIC_FORM_CONTROL_TYPE_DATEPICKER:
+                    formModel.push(new DynamicDatepickerModel(model, model["cls"]));
                     break;
 
                 case DYNAMIC_FORM_CONTROL_TYPE_GROUP: