+++ CHANGELOG.md
@@ -1,8 +1,18 @@
+# 1.3.10
+
+### **Features** 
+
+* initial addition of date control models
+* initial support for `Calendar` in `ui-kendo` and `ui-primeng`
+* `fromJSON(json: string | Object[])` function refactored to accept raw JSON strings and to revive `Date` objects
+* update to Angular `2.4.5`
+
+
 # 1.3.9
 
 ### **Features** 
 
-* functions to add and remove form controls after initialization added in `DynamicFormService` (closes [#252](https://github.com/udos86/ng2-dynamic-forms/issues/247))
+* functions to add and remove form controls after initialization added in `DynamicFormService` (closes [#252](https://github.com/udos86/ng2-dynamic-forms/issues/252))
 * initial support for `AutoComplete`, `MaskedTextBox` and `NumericTextBox` in `ui-kendo`
 * update to Angular `2.4.4`